requirements.txt
torch==2.4.1
transformers==4.46.3
numpy==1.24.3
scikit-learn==1.3.2
tqdm==4.67.1
pandas==2.0.3
scipy==1.10.1
matplotlib-base==3.7.2
seaborn==0.13.2
pyyaml==6.0.2
torchvision>=0.10.0
pytorch-lightning>=2.0.0
python-dotenv>=0.19.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
tokenizers==0.20.3
huggingface-hub==0.26.2
pillow==10.4.0
requests==2.32.3
joblib==1.4.2
packaging==24.2